more precise unsigned types, warning fix(es)

+++ libexp/include/Galois/Runtime/dGraph_edgeCut.h
@@ -125,7 +125,7 @@ class hGraph_edgeCut : public hGraph<NodeTy, EdgeTy, BSPNode, BSPEdge> {
 
       uint64_t numNodes_to_divide = 0;
       if (isBipartite) {
-    	  for (unsigned int n = 0; n < g.size(); ++n){
+    	  for (uint64_t n = 0; n < g.size(); ++n){
     		  if(std::distance(g.edge_begin(n), g.edge_end(n))){
                   ++numNodes_to_divide;
                   last_nodeID_withEdges_bipartite = n;
